np.arange(5).astype(np.float)

+++ experiments/supervised/dcase_script.py
@@ -4,6 +4,7 @@ import joblib
 import numpy as np
 import pickle
 import h5py
+import numpy as np
 import os.path
 import scipy.stats
 import sklearn.decomposition
@@ -20,7 +21,7 @@ np.set_printoptions(precision=2)
 
 # Evaluate role of compression
 exp1 = []
-for C in np.arange(5):
+for C in np.arange(5).astype(np.float):
     exp1.append(dcase_svm(0, 12, False, "time", "none", "early", C))
     exp1.append(dcase_svm(0, 12, False, "time", "log", "early", C))
     exp1.append(dcase_svm(0, 12, False, "time", "logmedian", "early", C))
@@ -30,7 +31,7 @@ for C in np.arange(5):
 
 # Compare time vs. time-frequency scattering, and role of data augmentation
 exp2 = []
-for C in np.arange(5):
+for C in np.arange(5).astype(np.float):
     exp2.append(dcase_svm(0, 12, False, "time", "log", "early", C))
     exp2.append(dcase_svm(0, 12, False, "timefrequency", "log", "early", C))
     exp2.append(dcase_svm(0, 12, False, "time", "log", "late", C))
